refactor(omegaform): extract defaultsValueFromSchema
